# backend/docs/swagger.yaml
definitions:
...
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/nat-config:
    get:
      consumes:
      - application/json
      description: 获取NAT配置
      parameters:
      - description: 'IP版本: v4 或 v6'
        in: query
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取NAT配置
      tags:
      - Admin API - NAT配置
    post:
      consumes:
      - application/json
      description: 保存NAT配置
      parameters:
      - description: 'IP版本: v4 或 v6'
        in: query
        name: version
        required: true
        type: string
      - description: NAT配置
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 保存失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 保存NAT配置
      tags:
      - Admin API - NAT配置
...
swagger: "2.0"